modified app, config,database and model
